photo-sharing-server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
distlib==0.3.6
filelock==3.12.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.8.6
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pipenv==2023.6.2
platformdirs==3.5.1
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
Werkzeug==2.3.6
python-dotenv==1.0.0
google-cloud-vision==3.5.0